asoc: remove overwriting for channel map control name
In kernel5.15, there is a hash for each control and it's calculated
with control name, iface and device before adding it to alsa ctl list.
So we cannoot overwrite ctl name after it's added in alsa list. Otherwise,
the inconsistency between the saved hash and the hash calculated during
use will result in control execution failure.

Change-Id: I2be4a1b0622838b62325d29cbf8dc46ef7293790
Signed-off-by: Deru Wang <quic_deruwang@quicinc.com>