ASoC: msm-pcm-q6-v2: Add dsp buf check
Current logic copies user buf size of data
from the avail dsp buf at a given offset.
If this offset returned from DSP in READ_DONE event
goes out of bounds or is corrupted, then it can lead to
out of bounds DSP buffer access, resulting in memory fault.
Fix is to add check for this buf offset, if it is within
the buf size range.

Change-Id: Iea2a7224cb0d3e33e5b5de3c593a4b46d1aa1fa8
Signed-off-by: Ramireddy KrishnaKanth Reddy <quic_ramikris@quicinc.com>