asoc: kona: Add tx macro swr use case.
It was found that device is crashing when swr port param is
being used for TX macro. This is due to modified usecase value
for which swr port params are not present in swr port map. so added
new usecase for TX macro with same port param in swr port map.

Change-Id: I101285edb3b5f6b0266a818d61933c3d9344c15b
Signed-off-by: Harendra Gautam <quic_hgautam@quicinc.com>