dsp: q6voice: Add check for DSP payload size
there is no check for payload size for the
response received for VSS related cmds, but
the payload is being accessed using two uint32_t
pointer variables. It is needed to ensure that
the payload is of size at least 2*sizeof(uint32_t)
to avoid over reading the payload pointer.
this fix is to ensure the above said check.

Change-Id: I54767add58f11a8696bb78efe51e18eab4c7b0d3
Signed-off-by: Srinivas Marka <quic_smarka@quicinc.com>