asoc: remove transcode loopback front end file
Remove msm-transcode-loopback-q6-v2.c as it is not added to Kbuild and
is currently unused. The security team reported a UAF issue in this file,
but since it is not compiled or used, we cannot validate or fix the issue.
Therefore, we are removing the file to clean up unused code.

No changes were made to the copyright content. This is a full file removal
and follows go/DriverSnapshot guidelines for legacy code cleanup.

Change-Id: I58e094503b45099f296c818197572de9d4c44a0e
Signed-off-by: Shalini Manjunatha <c_shalma@qti.qualcomm.com>
Signed-off-by: Renfei Qian <quic_renfqian@quicinc.com>